%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.data.frame.CBS}
\alias{as.data.frame.CBS}
\alias{CBS.as.data.frame}
\alias{as.data.frame,CBS-method}

\title{Gets the table of segments}

\description{
 Gets the table of segments.
}

\usage{
\method{as.data.frame}{CBS}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{data.frame}}, where each row corresponds to
  a unique segment.
}

\author{Henrik Bengtsson}

\seealso{
  Utilizes \code{\link[PSCBS:getSegments.CBS]{*getSegments}()}.
  For more information see \code{\link{CBS}}..
}


\keyword{internal}
\keyword{methods}
