%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{as.data.frame.PairedPSCBS}
\alias{as.data.frame.PairedPSCBS}
\alias{PairedPSCBS.as.data.frame}
\alias{as.data.frame.PairedPSCBS}
\alias{as.data.frame,PairedPSCBS-method}


 \title{Gets the table of segments}

 \description{
  Gets the table of segments.
 }

 \usage{\method{as.data.frame}{PairedPSCBS}(x, ...)}

 \arguments{
   \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{data.frame}}, where each row corresponds to
   a unique segment.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
