%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  exampleData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{exampleData}
\alias{exampleData.default}
\alias{exampleData}

\title{Gets an example data set}

\description{
 Gets an example data set.
}

\usage{
\method{exampleData}{default}(name=c("paired.chr01"), ...)
}

\arguments{
  \item{name}{A \code{\link[base]{character}} string specifying the name of the data set.}
  \item{...}{Not used.}
}

\value{
  Returns \code{\link[base]{data.frame}}.
}

\author{Henrik Bengtsson}


\keyword{IO}
