%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCBS.EXTS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{extractMinorMajorCNs.PairedPSCBS}
\alias{extractMinorMajorCNs.PairedPSCBS}
\alias{PairedPSCBS.extractMinorMajorCNs}
\alias{extractMinorMajorCNs,PairedPSCBS-method}
\alias{PairedPSCBS.extractC1C2}
\alias{extractC1C2.PairedPSCBS}
\alias{extractC1C2,PairedPSCBS-method}


\title{Extract minor and major copy-number mean levels per segment}

\description{
  Extract minor and major copy-number mean levels per segment.
}

\usage{
\method{extractMinorMajorCNs}{PairedPSCBS}(fit, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{data.frame}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[PSCBS:extractTCNAndDHs.PairedPSCBS]{*extractTCNAndDHs}()}
  For more information see \code{\link{PairedPSCBS}}.
}
\keyword{internal}
\keyword{methods}
