%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CBS.EXTS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{extractSegmentMeansByLocus.CBS}
\alias{extractSegmentMeansByLocus.CBS}
\alias{CBS.extractSegmentMeansByLocus}
\alias{extractSegmentMeansByLocus,CBS-method}

\title{Extracts segments means at each locus}

\description{
 Extracts segments means at each locus.
}

\usage{
\method{extractSegmentMeansByLocus}{CBS}(fit, ...)
}

\arguments{
 \item{...}{Arguments passed to \code{\link[PSCBS:getLocusData.CBS]{*getLocusData}()}.}
}

\value{
 Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of length \code{nbrOfLoci()}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{CBS}}.
}


\keyword{internal}
\keyword{methods}
