%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCBS.EXTS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


 \name{extractTCNAndDHs.PairedPSCBS}
\alias{extractTCNAndDHs.PairedPSCBS}
\alias{PairedPSCBS.extractTCNAndDHs}
\alias{extractTCNAndDHs,PairedPSCBS-method}

 \title{Extract TCN and DH mean levels per segment}

 \description{
   Extract TCN and DH mean levels per segment.
 }

 \usage{
\method{extractTCNAndDHs}{PairedPSCBS}(fit, ...)
}

 \arguments{
  \item{...}{Arguments passed to \code{getSegments()}.}
 }

 \value{
   Returns a \code{\link[base]{data.frame}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[PSCBS:extractMinorMajorCNs.PairedPSCBS]{*extractMinorMajorCNs}()}.
   For more information see \code{\link{PairedPSCBS}}.
 }
\keyword{internal}
\keyword{methods}
