%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getChromosomes.AbstractCBS}
\alias{getChromosomes.AbstractCBS}
\alias{AbstractCBS.getChromosomes}
\alias{getChromosomes,AbstractCBS-method}

 \title{Gets the set of chromosomes}

 \description{
   Gets the set of chromosomes in the segmentation result.
 }

 \usage{
\method{getChromosomes}{AbstractCBS}(this, ...)
}

 \arguments{
  \item{...}{Arguments passed to \code{\link[PSCBS:getSegments.AbstractCBS]{*getSegments}()}.}
 }

 \value{
   Returns a unique and sorted \code{\link[base]{vector}} of chromosomes segmented.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[PSCBS:nbrOfChromosomes.AbstractCBS]{*nbrOfChromosomes}()}.
   For more information see \code{\link{AbstractCBS}}.
 }
\keyword{internal}
\keyword{methods}
