%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCNData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getChromosomes.PairedPSCNData}
\alias{getChromosomes.PairedPSCNData}
\alias{PairedPSCNData.getChromosomes}
\alias{getChromosomes.PairedPSCNData}
\alias{getChromosomes,PairedPSCNData-method}


 \title{Gets the set of chromosomes}

 \description{
   Gets the set of chromosomes.
 }

 \usage{\method{getChromosomes}{PairedPSCNData}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns a unique and sorted \code{\link[base]{vector}} of chromosome indices.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[PSCBS:nbrOfChromosomes.PairedPSCNData]{*nbrOfChromosomes}()}.
   For more information see \code{\link{PairedPSCNData}}.
 }
\keyword{internal}
\keyword{methods}
