%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getSampleName.AbstractCBS}
\alias{getSampleName.AbstractCBS}
\alias{AbstractCBS.getSampleName}
\alias{getSampleName,AbstractCBS-method}
 \alias{AbstractCBS.sampleName}
\alias{sampleName.AbstractCBS}
\alias{sampleName,AbstractCBS-method}


 \title{Gets the name of the sample segmented}

 \description{
  Gets the name of the sample segmented.
 }

 \usage{
\method{getSampleName}{AbstractCBS}(fit, ...)
}

 \arguments{
   \item{...}{Not used.}
 }

 \value{
   Returns a \code{\link[base]{character}} string.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[PSCBS:setSampleName.AbstractCBS]{*setSampleName}()}.
   For more information see \code{\link{AbstractCBS}}..
 }
\keyword{internal}
\keyword{methods}
