%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CBS.CALL.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{mergeNonCalledSegments.CBS}
\alias{mergeNonCalledSegments.CBS}
\alias{CBS.mergeNonCalledSegments}
\alias{mergeNonCalledSegments,CBS-method}

 \title{Merge neighboring segments that are not called}

 \description{
   Merge neighboring segments that are not called
 }

 \usage{
\method{mergeNonCalledSegments}{CBS}(fit, ..., verbose=FALSE)
}

 \arguments{
  \item{...}{Not used.}
  \item{verbose}{\code{\link[R.utils]{Verbose}}.}
 }

 \value{
   Returns an object of the same class
   with the same of fewer number of segments.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{CBS}}.
 }


\keyword{internal}
\keyword{methods}
