%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfChromosomes.AbstractCBS}
\alias{nbrOfChromosomes.AbstractCBS}
\alias{AbstractCBS.nbrOfChromosomes}
\alias{nbrOfChromosomes.AbstractCBS}
\alias{nbrOfChromosomes,AbstractCBS-method}


 \title{Gets the number of chromosomes}

 \description{
   Gets the number of chromosomes.
 }

 \usage{\method{nbrOfChromosomes}{AbstractCBS}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[PSCBS:getChromosomes.AbstractCBS]{*getChromosomes}()}.
   For more information see \code{\link{AbstractCBS}}.
 }
\keyword{internal}
\keyword{methods}
