%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCNData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfChromosomes.PairedPSCNData}
\alias{nbrOfChromosomes.PairedPSCNData}
\alias{PairedPSCNData.nbrOfChromosomes}
\alias{nbrOfChromosomes.PairedPSCNData}
\alias{nbrOfChromosomes,PairedPSCNData-method}


 \title{Gets the number of chromosomes}

 \description{
   Gets the number of chromosomes.
 }

 \usage{\method{nbrOfChromosomes}{PairedPSCNData}(this, ...)}

 \arguments{
  \item{...}{Arguments passed to \code{\link[PSCBS:getChromosomes.PairedPSCNData]{*getChromosomes}()}.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   \code{\link[PSCBS:getChromosomes.PairedPSCNData]{*getChromosomes}()}.
   For more information see \code{\link{PairedPSCNData}}.
 }
\keyword{internal}
\keyword{methods}
