%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfLoci.AbstractCBS}
\alias{nbrOfLoci.AbstractCBS}
\alias{AbstractCBS.nbrOfLoci}
\alias{nbrOfLoci.AbstractCBS}
\alias{nbrOfLoci,AbstractCBS-method}


 \title{Gets the number of loci}

 \description{
   Gets the number of loci.
 }

 \usage{\method{nbrOfLoci}{AbstractCBS}(fit, splitters=FALSE, ...)}

 \arguments{
  \item{splitters, ...}{Arguments passed to \code{\link[PSCBS:getLocusData.AbstractCBS]{*getLocusData}()}.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{AbstractCBS}}.
 }
\keyword{internal}
\keyword{methods}
