%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  PairedPSCNData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfLoci.PairedPSCNData}
\alias{nbrOfLoci.PairedPSCNData}
\alias{PairedPSCNData.nbrOfLoci}
\alias{nbrOfLoci.PairedPSCNData}
\alias{nbrOfLoci,PairedPSCNData-method}


 \title{Gets the number of loci}

 \description{
   Gets the number of loci.
 }

 \usage{\method{nbrOfLoci}{PairedPSCNData}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{PairedPSCNData}}.
 }
\keyword{internal}
\keyword{methods}
