%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfSegments.AbstractCBS}
\alias{nbrOfSegments.AbstractCBS}
\alias{AbstractCBS.nbrOfSegments}
\alias{nbrOfSegments,AbstractCBS-method}

 \title{Gets the number of segments}

 \description{
   Gets the number of segments.
 }

 \usage{
\method{nbrOfSegments}{AbstractCBS}(this, splitters=FALSE, ...)
}

 \arguments{
  \item{splitters, ...}{Arguments passed to \code{\link[PSCBS:getSegments.AbstractCBS]{*getSegments}()}.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson}

 \seealso{
   \code{\link[PSCBS:nbrOfChangePoints.AbstractCBS]{*nbrOfChangePoints}()}
   \code{\link[PSCBS:nbrOfChromosomes.AbstractCBS]{*nbrOfChromosomes}()}
   For more information see \code{\link{AbstractCBS}}.
 }
\keyword{internal}
\keyword{methods}
