%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{normalizeTotalCNs.AbstractCBS}
\alias{normalizeTotalCNs.AbstractCBS}
\alias{AbstractCBS.normalizeTotalCNs}
\alias{normalizeTotalCNs,AbstractCBS-method}
\alias{normalizeTotalCNs}
\alias{normalizeTotalCNs.PSCBS}

\title{Normalizes copy numbers such that the whole-genome average total copy number is two}

\description{
 Normalizes copy numbers such that the whole-genome average total copy number is two.
}

\usage{
\method{normalizeTotalCNs}{AbstractCBS}(...)
}

\arguments{
  \item{...}{Additional arguments passed to the normalization method.}
}

\value{
  Returns a normalized AbstractCBS object of the same class as \code{fit}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AbstractCBS}}..
}
\keyword{internal}
\keyword{methods}
