%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.PLOT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{plotTracks.AbstractCBS}
\alias{plotTracks.AbstractCBS}
\alias{AbstractCBS.plotTracks}
\alias{plotTracks,AbstractCBS-method}

\title{Plots the segmentation result along the genome}

\description{
  Plots the segmentation result along the genome.
}

\usage{
\method{plotTracks}{AbstractCBS}(...)
}

\arguments{
 \item{...}{...}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AbstractCBS}}.
}
\keyword{internal}
\keyword{methods}
