%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{psCBS-package}
\alias{psCBS-package}
\alias{psCBS}
\docType{package}

\title{Package psCBS}


 \description{
   This package provides a method for robustly fitting a K-dimensional simplex in M dimensions..

   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires external packages.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{http://r-forge.r-project.org/}.
 }

 \section{To get started}{
   To get started, see:
   \enumerate{
     \item \code{\link{segmentByPairedPSCBS}}()
   }
 }

 \section{How to cite this package}{
  TBA.
 }

 \author{
  TBA.
 }

 \section{License}{
  TBA.
 }

 \references{
  TBA.
 }
\keyword{package}
