%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{save.AbstractCBS}
\alias{save.AbstractCBS}
\alias{AbstractCBS.save}
\alias{save,AbstractCBS-method}

 \title{Saves an AbstractCBS object to file}

 \description{
  Saves an AbstractCBS object to file.
 }

 \usage{
\method{save}{AbstractCBS}(this, ...)
}

 \arguments{
   \item{...}{Additional arguments passed to \code{\link[R.utils]{saveObject}}.}
 }

 \value{
   Returns what \code{\link[R.utils]{saveObject}} returns.
 }

 \author{Henrik Bengtsson}

 \seealso{
   Internally \code{\link[R.utils]{saveObject}} is used.
   To load an object, see \code{\link[PSCBS:load.AbstractCBS]{*load}()}.
   For more information see \code{\link{AbstractCBS}}..
 }
\keyword{internal}
\keyword{methods}
