%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{setSampleName.AbstractCBS}
\alias{setSampleName.AbstractCBS}
\alias{AbstractCBS.setSampleName}
\alias{setSampleName,AbstractCBS-method}
 \alias{AbstractCBS.sampleName<-}
\alias{sampleName<-.AbstractCBS}
\alias{sampleName<-,AbstractCBS-method}


 \title{Sets the name of the sample segmented}

 \description{
  Sets the name of the sample segmented.
 }

 \usage{
\method{setSampleName}{AbstractCBS}(fit, name, ...)
}

 \arguments{
   \item{name}{A \code{\link[base]{character}} string.}
   \item{...}{Not used.}
 }

 \value{
   Returns (invisibly) an updated object.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{AbstractCBS}}..
 }


\keyword{internal}
\keyword{methods}
