%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{updateMeans.AbstractCBS}
\alias{updateMeans.AbstractCBS}
\alias{AbstractCBS.updateMeans}
\alias{updateMeans.AbstractCBS}
\alias{updateMeans,AbstractCBS-method}

\alias{updateMeans.CBS}
\alias{updateMeans.NonPairedPSCBS}
\alias{updateMeans.PairedPSCBS}

\title{Updates the CN mean levels for each segment independently}

\description{
 Updates the CN mean levels for each segment independently as if they were one large segment.
 The locus-level data is not updated/modified.
}

\usage{\method{updateMeans}{AbstractCBS}(...)}

\arguments{
 \item{...}{Arguments specific to the class.}
}

\value{
  Returns an object of the same class.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
