%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CBS.EXTS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{extractSegmentMeansByLocus.CBS}
\alias{extractSegmentMeansByLocus.CBS}
\alias{CBS.extractSegmentMeansByLocus}
\alias{extractSegmentMeansByLocus.CBS}
\alias{extractSegmentMeansByLocus,CBS-method}


\title{Extracts segments means at each locus}

\description{
 Extracts segments means at each locus.
}

\usage{\method{extractSegmentMeansByLocus}{CBS}(fit, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of length \code{\link[PSCBS:nbrOfLoci.CBS]{*nbrOfLoci}()}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CBS}}.
}


\keyword{internal}
\keyword{methods}
