%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AbstractCBS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{nbrOfSegments.AbstractCBS}
\alias{nbrOfSegments.AbstractCBS}
\alias{AbstractCBS.nbrOfSegments}
\alias{nbrOfSegments.AbstractCBS}
\alias{nbrOfSegments,AbstractCBS-method}


 \title{Gets the number of segments}

 \description{
   Gets the number of segments.
 }

 \usage{\method{nbrOfSegments}{AbstractCBS}(this, splitters=FALSE, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{AbstractCBS}}.
 }
\keyword{internal}
\keyword{methods}
