%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{append}
\alias{append.PSCBS}
\alias{append.default}
\alias{arrowsC1C2}
\alias{arrowsC1C2.PairedPSCBS}
\alias{arrowsDeltaC1C2}
\alias{arrowsDeltaC1C2.PairedPSCBS}
\alias{bootstrapCIs}
\alias{bootstrapCIs.PairedPSCBS}
\alias{bootstrapDHByRegion}
\alias{bootstrapDHByRegion.PairedPSCBS}
\alias{callABandHighAI}
\alias{callABandHighAI.PairedPSCBS}
\alias{callExtremeAllelicImbalance}
\alias{callExtremeAllelicImbalance.PairedPSCBS}
\alias{callSegments}
\alias{callSegments.PairedPSCBS}
\alias{drawLevels}
\alias{drawLevels.DNAcopy}
\alias{drawLevels.PairedPSCBS}
\alias{extractByChromosome}
\alias{extractByChromosome.PSCBS}
\alias{extractC1C2}
\alias{extractC1C2.PairedPSCBS}
\alias{extractDeltaC1C2}
\alias{extractDeltaC1C2.PairedPSCBS}
\alias{extractMinorMajorCNs}
\alias{extractMinorMajorCNs.PairedPSCBS}
\alias{extractTCNAndDHs}
\alias{extractTCNAndDHs.PairedPSCBS}
\alias{getChromosomes}
\alias{getChromosomes.PSCBS}
\alias{linesC1C2}
\alias{linesC1C2.PairedPSCBS}
\alias{linesDeltaC1C2}
\alias{linesDeltaC1C2.PairedPSCBS}
\alias{nbrOfChromosomes}
\alias{nbrOfChromosomes.PSCBS}
\alias{nbrOfSegments}
\alias{nbrOfSegments.PSCBS}
\alias{plot.PairedPSCBS}
\alias{plotC1C2}
\alias{plotC1C2.PairedPSCBS}
\alias{plotDeltaC1C2}
\alias{plotDeltaC1C2.PairedPSCBS}
\alias{pointsC1C2}
\alias{pointsC1C2.PairedPSCBS}
\alias{pointsDeltaC1C2}
\alias{pointsDeltaC1C2.PairedPSCBS}
\alias{postsegmentTCN}
\alias{postsegmentTCN.PairedPSCBS}
\alias{print.PairedPSCBS}
\alias{subsetByDhSegments}
\alias{subsetByDhSegments.PairedPSCBS}
\alias{plotTracksManyChromosomes}
\alias{plotTracksManyChromosomes.PairedPSCBS}
\alias{tileChromosomes}
\alias{tileChromosomes.PairedPSCBS}
\alias{callExtremeAllelicImbalanceByDH}
\alias{callExtremeAllelicImbalanceByDH.PairedPSCBS}
\alias{as.data.frame.CBS}
\alias{drawLevels.CBS}
\alias{getChromosomes.CBS}
\alias{nbrOfChromosomes.CBS}
\alias{plot.CBS}
\alias{plotTracksManyChromosomes.CBS}
\alias{tileChromosomes.CBS}
\alias{bootstrapTCNandDHByRegion}
\alias{bootstrapTCNandDHByRegion.PairedPSCBS}
\alias{joinSegments}
\alias{joinSegments.CBS}
\alias{drawConfidenceBands}
\alias{drawConfidenceBands.PairedPSCBS}
\alias{extractByChromosomes}
\alias{extractByChromosomes.PSCBS}
\alias{dropSegmentationOutliers}
\alias{dropSegmentationOutliers.default}
\alias{callABandLowC1}
\alias{callABandLowC1.PairedPSCBS}
\alias{callLowC1ByC1}
\alias{callLowC1ByC1.PairedPSCBS}
\alias{estimateStdDevForHeterozygousBAF}
\alias{estimateStdDevForHeterozygousBAF.PairedPSCBS}
\alias{extractByRegion}
\alias{extractByRegion.PairedPSCBS}
\alias{extractByRegions}
\alias{extractByRegions.PairedPSCBS}
\alias{mergeTwoSegments}
\alias{mergeTwoSegments.PairedPSCBS}
\alias{updateMeans}
\alias{updateMeans.PairedPSCBS}
\alias{dropByRegions}
\alias{dropByRegions.PairedPSCBS}
\alias{extractLocusLevelC1C2}
\alias{extractLocusLevelC1C2.PairedPSCBS}
\alias{extractLocusLevelTCN}
\alias{extractLocusLevelTCN.PairedPSCBS}
\alias{plotTracks2}
\alias{plotTracks2.PairedPSCBS}
\alias{estimateMeanForDH}
\alias{estimateMeanForDH.PairedPSCBS}
\alias{applyByRegion}
\alias{applyByRegion.PairedPSCBS}
\alias{estimateHighDHQuantileAtAB}
\alias{estimateHighDHQuantileAtAB.PairedPSCBS}
\alias{estimateKappa}
\alias{estimateKappaByC1Density}
\alias{callAllelicBalance}
\alias{callAllelicBalanceByDH}
\alias{estimateDeltaAB}
\alias{estimateDeltaABBySmallDH}
\alias{estimateDeltaLOH}
\alias{estimateDeltaLOHByMinC1ForNonAB}
\alias{callAB}
\alias{callAllelicBalance.default}
\alias{callLOH}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
