%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{append}
\alias{append.default}
\alias{arrowsC1C2}
\alias{arrowsC1C2.PairedPSCBS}
\alias{arrowsDeltaC1C2}
\alias{arrowsDeltaC1C2.PairedPSCBS}
\alias{bootstrapCIs}
\alias{bootstrapCIs.PairedPSCBS}
\alias{bootstrapDHByRegion}
\alias{bootstrapDHByRegion.PairedPSCBS}
\alias{callABandHighAI}
\alias{callABandHighAI.PairedPSCBS}
\alias{callExtremeAllelicImbalance}
\alias{callExtremeAllelicImbalance.PairedPSCBS}
\alias{drawLevels}
\alias{drawLevels.DNAcopy}
\alias{drawLevels.PairedPSCBS}
\alias{extractByChromosome}
\alias{extractByChromosome.PSCBS}
\alias{extractC1C2}
\alias{extractDeltaC1C2}
\alias{extractDeltaC1C2.PairedPSCBS}
\alias{extractMinorMajorCNs}
\alias{extractTCNAndDHs}
\alias{getChromosomes}
\alias{linesC1C2}
\alias{linesC1C2.PairedPSCBS}
\alias{linesDeltaC1C2}
\alias{linesDeltaC1C2.PairedPSCBS}
\alias{nbrOfChromosomes}
\alias{nbrOfSegments}
\alias{plot.PairedPSCBS}
\alias{plotC1C2}
\alias{plotC1C2.PairedPSCBS}
\alias{plotDeltaC1C2}
\alias{plotDeltaC1C2.PairedPSCBS}
\alias{pointsC1C2}
\alias{pointsC1C2.PairedPSCBS}
\alias{pointsDeltaC1C2}
\alias{pointsDeltaC1C2.PairedPSCBS}
\alias{postsegmentTCN}
\alias{postsegmentTCN.PairedPSCBS}
\alias{print.PairedPSCBS}
\alias{subsetByDhSegments}
\alias{subsetByDhSegments.PairedPSCBS}
\alias{plotTracksManyChromosomes}
\alias{plotTracksManyChromosomes.PairedPSCBS}
\alias{tileChromosomes}
\alias{tileChromosomes.PairedPSCBS}
\alias{callExtremeAllelicImbalanceByDH}
\alias{callExtremeAllelicImbalanceByDH.PairedPSCBS}
\alias{drawLevels.CBS}
\alias{getChromosomes.CBS}
\alias{nbrOfChromosomes.CBS}
\alias{plot.CBS}
\alias{plotTracksManyChromosomes.CBS}
\alias{tileChromosomes.CBS}
\alias{bootstrapTCNandDHByRegion}
\alias{bootstrapTCNandDHByRegion.PairedPSCBS}
\alias{joinSegments}
\alias{joinSegments.CBS}
\alias{drawConfidenceBands}
\alias{drawConfidenceBands.PairedPSCBS}
\alias{extractByChromosomes}
\alias{extractByChromosomes.PSCBS}
\alias{dropSegmentationOutliers}
\alias{dropSegmentationOutliers.default}
\alias{callABandLowC1}
\alias{callABandLowC1.PairedPSCBS}
\alias{callLowC1ByC1}
\alias{callLowC1ByC1.PairedPSCBS}
\alias{estimateStdDevForHeterozygousBAF}
\alias{estimateStdDevForHeterozygousBAF.PairedPSCBS}
\alias{extractByRegion}
\alias{extractByRegion.PairedPSCBS}
\alias{extractByRegions}
\alias{extractByRegions.PairedPSCBS}
\alias{mergeTwoSegments}
\alias{updateMeans}
\alias{updateMeans.PairedPSCBS}
\alias{dropByRegions}
\alias{extractLocusLevelC1C2}
\alias{extractLocusLevelC1C2.PairedPSCBS}
\alias{extractLocusLevelTCN}
\alias{extractLocusLevelTCN.PairedPSCBS}
\alias{plotTracks2}
\alias{plotTracks2.PairedPSCBS}
\alias{estimateMeanForDH}
\alias{estimateMeanForDH.PairedPSCBS}
\alias{applyByRegion}
\alias{applyByRegion.PairedPSCBS}
\alias{estimateHighDHQuantileAtAB}
\alias{estimateHighDHQuantileAtAB.PairedPSCBS}
\alias{estimateKappa}
\alias{estimateKappaByC1Density}
\alias{callAllelicBalance}
\alias{callAllelicBalanceByDH}
\alias{estimateDeltaAB}
\alias{estimateDeltaABBySmallDH}
\alias{estimateDeltaLOH}
\alias{estimateDeltaLOHByMinC1ForNonAB}
\alias{callAB}
\alias{callAllelicBalance.default}
\alias{callLOH}
\alias{getChromosomes.DNAcopy}
\alias{getSampleNames}
\alias{getSampleNames.CBS}
\alias{getSampleNames.DNAcopy}
\alias{nbrOfLoci}
\alias{nbrOfLoci.CBS}
\alias{nbrOfLoci.DNAcopy}
\alias{nbrOfSamples}
\alias{nbrOfSamples.DNAcopy}
\alias{nbrOfSegments.DNAcopy}
\alias{estimateStandardDeviation}
\alias{estimateStandardDeviation.DNAcopy}
\alias{extractSegmentMeansByLocus}
\alias{extractSegmentMeansByLocus.DNAcopy}
\alias{as.CBS}
\alias{as.DNAcopy}
\alias{as.character.CBS}
\alias{extractByChromosomes.CBS}
\alias{getLocusData}
\alias{getLocusData.CBS}
\alias{getSegments}
\alias{getSegments.CBS}
\alias{nbrOfSegments.CBS}
\alias{print.CBS}
\alias{subset.CBS}
\alias{callAmplifications}
\alias{callGainsAndLosses}
\alias{callOutliers}
\alias{getSampleName}
\alias{getSampleName.CBS}
\alias{sampleName}
\alias{sampleName.CBS}
\alias{sampleName<-}
\alias{sampleName<-.CBS}
\alias{extractCallsByLocus}
\alias{extractCallsByLocus.CBS}
\alias{getSignalType}
\alias{getSignalType.CBS}
\alias{signalType}
\alias{signalType.CBS}
\alias{signalType<-}
\alias{signalType<-.CBS}
\alias{writeLocusData}
\alias{writeLocusData.CBS}
\alias{writeSegments}
\alias{writeSegments.CBS}
\alias{getCallStatistics}
\alias{getCallStatistics.CBS}
\alias{getChromosomeRanges}
\alias{getChromosomeRanges.CBS}
\alias{getFGA}
\alias{getFGG}
\alias{getFGL}
\alias{getFractionOfGenomeAltered}
\alias{getFractionOfGenomeGained}
\alias{getFractionOfGenomeLost}
\alias{isWholeChromosomeGained}
\alias{isWholeChromosomeGained.CBS}
\alias{isWholeChromosomeLost}
\alias{isWholeChromosomeLost.CBS}
\alias{nbrOfAmplifications}
\alias{nbrOfAmplifications.CBS}
\alias{nbrOfGains}
\alias{nbrOfGains.CBS}
\alias{nbrOfLosses}
\alias{nbrOfLosses.CBS}
\alias{drawCentromeres}
\alias{drawCentromeres.CBS}
\alias{getChromosomeOffsets}
\alias{getChromosomeOffsets.CBS}
\alias{highlightCalls}
\alias{highlightCalls.CBS}
\alias{highlightLocusCalls}
\alias{highlightLocusCalls.CBS}
\alias{as.data.frame.PSCBS}
\alias{extractByChromosome.AbstractCBS}
\alias{extractByChromosomes.AbstractCBS}
\alias{print.AbstractCBS}
\alias{setSampleName}
\alias{setLocusData}
\alias{setLocusData.AbstractCBS}
\alias{setSegments}
\alias{setSegments.AbstractCBS}
\alias{getLocusData.PSCBS}
\alias{tileChromosomes.AbstractCBS}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
