%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{all.equal.AbstractCBS}
\alias{all.equal.CBS}
\alias{append}
\alias{append.default}
\alias{applyByRegion}
\alias{applyByRegion.PairedPSCBS}
\alias{arrowsC1C2}
\alias{arrowsC1C2.PairedPSCBS}
\alias{arrowsDeltaC1C2}
\alias{arrowsDeltaC1C2.PairedPSCBS}
\alias{as.CBS}
\alias{as.DNAcopy}
\alias{as.character.CBS}
\alias{as.data.frame.PSCBS}
\alias{bootstrapCIs}
\alias{bootstrapCIs.PairedPSCBS}
\alias{bootstrapDHByRegion}
\alias{bootstrapDHByRegion.PairedPSCBS}
\alias{bootstrapTCNandDHByRegion}
\alias{callAB}
\alias{callABandHighAI}
\alias{callABandHighAI.PairedPSCBS}
\alias{callABandLowC1}
\alias{callABandLowC1.PairedPSCBS}
\alias{callAllelicBalance}
\alias{callAllelicBalance.default}
\alias{callAllelicBalanceByDH}
\alias{callAmplifications}
\alias{callArms}
\alias{callArms.CBS}
\alias{callExtremeAllelicImbalanceByDH}
\alias{callExtremeAllelicImbalanceByDH.PairedPSCBS}
\alias{callGainsAndLosses}
\alias{callLOH}
\alias{callLowC1ByC1}
\alias{callLowC1ByC1.PairedPSCBS}
\alias{callOutliers}
\alias{drawCentromeres}
\alias{drawCentromeres.CBS}
\alias{drawChromosomes}
\alias{drawChromosomes.CBS}
\alias{drawConfidenceBands}
\alias{drawConfidenceBands.PairedPSCBS}
\alias{drawLevels}
\alias{drawLevels.CBS}
\alias{drawLevels.DNAcopy}
\alias{drawLevels.PairedPSCBS}
\alias{dropChangePoint}
\alias{dropRegion}
\alias{dropRegions}
\alias{dropSegmentationOutliers}
\alias{dropSegmentationOutliers.default}
\alias{estimateDeltaAB}
\alias{estimateDeltaABBySmallDH}
\alias{estimateDeltaLOH}
\alias{estimateDeltaLOHByMinC1ForNonAB}
\alias{estimateHighDHQuantileAtAB}
\alias{estimateHighDHQuantileAtAB.PairedPSCBS}
\alias{estimateKappa}
\alias{estimateKappaByC1Density}
\alias{estimateMeanForDH}
\alias{estimateMeanForDH.PairedPSCBS}
\alias{estimateStandardDeviation}
\alias{estimateStandardDeviation.DNAcopy}
\alias{estimateStdDevForHeterozygousBAF}
\alias{estimateStdDevForHeterozygousBAF.PairedPSCBS}
\alias{extractC1C2}
\alias{extractCallsByLocus}
\alias{extractCallsByLocus.PairedPSCBS}
\alias{extractCallsByLocus.CBS}
\alias{extractChromosome}
\alias{extractChromosome.AbstractCBS}
\alias{extractChromosomes}
\alias{extractChromosomes.AbstractCBS}
\alias{extractChromosomes.CBS}
\alias{extractChromosomes.PSCBS}
\alias{extractDeltaC1C2}
\alias{extractDeltaC1C2.PairedPSCBS}
\alias{extractLocusLevelC1C2}
\alias{extractLocusLevelC1C2.PairedPSCBS}
\alias{extractLocusLevelTCN}
\alias{extractLocusLevelTCN.PairedPSCBS}
\alias{extractMinorMajorCNs}
\alias{extractRegion}
\alias{extractRegion.AbstractCBS}
\alias{extractRegions}
\alias{extractRegions.AbstractCBS}
\alias{extractSegmentMeansByLocus}
\alias{extractSegmentMeansByLocus.DNAcopy}
\alias{extractTCNAndDHs}
\alias{getCallStatistics}
\alias{getCallStatisticsByArms}
\alias{getCallStatisticsByArms.CBS}
\alias{getChromosomeOffsets}
\alias{getChromosomeRanges}
\alias{getChromosomeRanges.CBS}
\alias{getChromosomes}
\alias{getChromosomes.DNAcopy}
\alias{getFGA}
\alias{getFGG}
\alias{getFGL}
\alias{getFractionOfGenomeAltered}
\alias{getFractionOfGenomeGained}
\alias{getFractionOfGenomeLost}
\alias{getLocusData}
\alias{getLocusData.CBS}
\alias{getLocusData.PSCBS}
\alias{getSampleName}
\alias{getSampleNames}
\alias{getSampleNames.DNAcopy}
\alias{getSegments}
\alias{getSegments.CBS}
\alias{getSignalType}
\alias{getSignalType.CBS}
\alias{highlightArmCalls}
\alias{highlightArmCalls.CBS}
\alias{highlightCalls}
\alias{highlightCalls.CBS}
\alias{highlightLocusCalls}
\alias{highlightLocusCalls.CBS}
\alias{isWholeChromosomeGained}
\alias{isWholeChromosomeGained.CBS}
\alias{isWholeChromosomeLost}
\alias{isWholeChromosomeLost.CBS}
\alias{joinSegments}
\alias{linesC1C2}
\alias{linesC1C2.PairedPSCBS}
\alias{linesDeltaC1C2}
\alias{linesDeltaC1C2.PairedPSCBS}
\alias{mergeNonCalledSegments}
\alias{mergeTwoSegments}
\alias{mergeTwoSegments.CBS}
\alias{nbrOfAmplifications}
\alias{nbrOfAmplifications.CBS}
\alias{nbrOfChangePoints}
\alias{nbrOfChromosomes}
\alias{nbrOfGains}
\alias{nbrOfGains.CBS}
\alias{nbrOfLoci}
\alias{nbrOfLoci.DNAcopy}
\alias{nbrOfLosses}
\alias{nbrOfLosses.CBS}
\alias{nbrOfSamples}
\alias{nbrOfSamples.DNAcopy}
\alias{nbrOfSegments}
\alias{nbrOfSegments.DNAcopy}
\alias{plot.CBS}
\alias{plot.PairedPSCBS}
\alias{plotC1C2}
\alias{plotC1C2.PairedPSCBS}
\alias{plotDeltaC1C2}
\alias{plotDeltaC1C2.PairedPSCBS}
\alias{plotTracksManyChromosomes.CBS}
\alias{pointsC1C2}
\alias{pointsC1C2.PairedPSCBS}
\alias{pointsDeltaC1C2}
\alias{pointsDeltaC1C2.PairedPSCBS}
\alias{postsegmentTCN}
\alias{postsegmentTCN.PairedPSCBS}
\alias{print.AbstractCBS}
\alias{sampleName}
\alias{sampleName<-}
\alias{setLocusData}
\alias{setSampleName}
\alias{setSegments}
\alias{signalType}
\alias{signalType.CBS}
\alias{signalType<-}
\alias{signalType<-.CBS}
\alias{subset.CBS}
\alias{tileChromosomes}
\alias{tileChromosomes.AbstractCBS}
\alias{tileChromosomes.CBS}
\alias{tileChromosomes.PairedPSCBS}
\alias{writeLocusData}
\alias{writeLocusData.CBS}
\alias{writeSegments}
\alias{extractSegments.AbstractCBS}
\alias{extractSegment.AbstractCBS}
\alias{extractSegment}
\alias{extractSegments}
\alias{extractSegments.CBS}
\alias{extractSegments.PairedPSCBS}
\alias{isSegmentSplitter}
\alias{isSegmentSplitter.CBS}
\alias{isSegmentSplitter.PSCBS}
\alias{extractCNs}
\alias{extractCNs.AbstractCBS}
\alias{extractCNs.CBS}
\alias{extractCNs.PairedPSCBS}
\alias{extractTotalCNs}
\alias{extractTotalCNs.CBS}
\alias{getSegmentSizes}
\alias{getSegmentSizes.AbstractCBS}
\alias{getSegmentSizes.CBS}
\alias{getSegmentSizes.PSCBS}
\alias{sampleCNs}
\alias{sampleCNs.AbstractCBS}
\alias{mergeThreeSegments}
\alias{callROH}
\alias{callROHOneSegment}
\alias{callROHOneSegment.PairedPSCBS}
\alias{testROH}
\alias{pruneBySdUndo}
\alias{updateBoundaries}
\alias{updateBoundaries.CBS}
\alias{resegment}
\alias{resegment.AbstractCBS}
\alias{resegment.CBS}
\alias{resegment.PairedPSCBS}
\alias{callSegmentationOutliers.data.frame}
\alias{dropSegmentationOutliers.data.frame}
\alias{hclustCNs}
\alias{pruneByHClust}
\alias{updateMeansTogether}
\alias{updateMeans}
\alias{drawChangePoints}
\alias{drawChangePoints.AbstractCBS}
\alias{drawKnownSegments}
\alias{drawKnownSegments.AbstractCBS}
\alias{drawChangePoints.CBS}
\alias{drawChangePoints.PSCBS}
\alias{drawChangePointsC1C2}
\alias{drawChangePointsC1C2.PairedPSCBS}
\alias{extractDhSegment}
\alias{extractDhSegment.PairedPSCBS}
\alias{callNTCN}
\alias{callCopyNeutral}
\alias{callCopyNeutralByTCNofAB}
\alias{dropChangePoints}
\alias{calcStatsForCopyNeutralABs}
\alias{calcStatsForCopyNeutralABs.PairedPSCBS}
\alias{callGainNeutralLoss}
\alias{callGNL}
\alias{renameChromosomes}
\alias{renameChromosomes.AbstractCBS}
\alias{report}
\alias{writeSegments.DNAcopy}
\alias{estimateDeltaCN}
\alias{estimateDeltaCN.PairedPSCBS}
\alias{shiftTCN}
\alias{shiftTCN.AbstractCBS}
\alias{shiftTCN.CBS}
\alias{shiftTCN.PairedPSCBS}
\alias{getChromosomeOffsets.AbstractCBS}
\alias{getChromosomeRanges.PairedPSCBS}
\alias{getChromosomeRanges.AbstractCBS}
\alias{seqOfSegmentsByDP}
\alias{seqOfSegmentsByDP.AbstractCBS}
\alias{seqOfSegmentsByDP.CBS}
\alias{seqOfSegmentsByDP.PairedPSCBS}
\alias{seqOfSegmentsByDP.matrix}
\alias{getChromosomeOffsets.PairedPSCBS}
\alias{getLocusSignalNames}
\alias{getLocusSignalNames.AbstractCBS}
\alias{getLocusSignalNames.CBS}
\alias{getLocusSignalNames.PSCBS}
\alias{getSegmentTrackPrefixes}
\alias{getSegmentTrackPrefixes.AbstractCBS}
\alias{getSegmentTrackPrefixes.CBS}
\alias{getSegmentTrackPrefixes.PSCBS}
\alias{pruneByDP}
\alias{getMeanEstimators}
\alias{getMeanEstimators.AbstractCBS}
\alias{resegment.NonPairedPSCBS}
\alias{setMeanEstimators}
\alias{setMeanEstimators.AbstractCBS}
\alias{resetSegments}
\alias{getLocusData.NonPairedPSCBS}
\alias{getLocusData.PairedPSCBS}
\alias{plotTracks}
\alias{plotTracks1}
\alias{plotTracks2}
\alias{plotTracksManyChromosomes}
\alias{getSmoothLocusData}
\alias{bootstrapSegmentsAndChangepoints}
\alias{bootstrapSegmentsAndChangepoints.PairedPSCBS}
\alias{getChangePoints}
\alias{getChangePoints.AbstractCBS}
\alias{getChangePoints.CBS}
\alias{getChangePoints.PSCBS}
\alias{clearBootstrapSummaries}
\alias{clearBootstrapSummaries.PairedPSCBS}
\alias{findBootstrapSummaries}
\alias{findBootstrapSummaries.PairedPSCBS}
\alias{hasBootstrapSummaries}
\alias{hasBootstrapSummaries.PairedPSCBS}
\alias{clearCalls}
\alias{clearCalls.AbstractCBS}
\alias{estimateDeltaCN.CBS}
\alias{extractSegmentDataByLocus}
\alias{extractSegmentDataByLocus.PairedPSCBS}
\alias{updateMeansC1C2}
\alias{updateMeansC1C2.PairedPSCBS}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson}


\keyword{documentation}
\keyword{internal}
