%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{nbrOfLoci}
\alias{nbrOfLoci.psCNA}
\alias{nbrOfReferenceSamples}
\alias{nbrOfReferenceSamples.psCNA}
\alias{getSampleNames}
\alias{getSampleNames.psCNA}
\alias{nbrOfSamples}
\alias{nbrOfSamples.psCNA}
\alias{psSegmentPaired}
\alias{callSegmentsByPSCBS}
\alias{callSegmentsByPSCBS.default}
\alias{drawLevels}
\alias{drawLevels.DNAcopy}
\alias{as.data.frame.PairedPSCBS}
\alias{bootstrapCIs}
\alias{bootstrapCIs.PairedPSCBS}
\alias{callSegments}
\alias{callSegments.PairedPSCBS}
\alias{drawLevels.PairedPSCBS}
\alias{plot.PairedPSCBS}
\alias{print.PairedPSCBS}
\alias{extractMinorMajorCNs}
\alias{extractMinorMajorCNs.PairedPSCBS}
\alias{extractTotalAndDH}
\alias{extractTotalAndDH.PairedPSCBS}
\alias{plotC1C2}
\alias{plotC1C2.PairedPSCBS}
\alias{pointsC1C2}
\alias{pointsC1C2.PairedPSCBS}
\alias{extractTCNAndDHs}
\alias{extractTCNAndDHs.PairedPSCBS}
\alias{arrowsC1C2}
\alias{arrowsC1C2.PairedPSCBS}
\alias{linesC1C2}
\alias{linesC1C2.PairedPSCBS}
\alias{subsetBySegments}
\alias{subsetBySegments.PairedPSCBS}
\alias{bootstrapDHByRegion}
\alias{bootstrapDHByRegion.PairedPSCBS}
\alias{arrowsDeltaC1C2}
\alias{arrowsDeltaC1C2.PairedPSCBS}
\alias{extractC1C2}
\alias{extractC1C2.PairedPSCBS}
\alias{extractDeltaC1C2}
\alias{extractDeltaC1C2.PairedPSCBS}
\alias{linesDeltaC1C2}
\alias{linesDeltaC1C2.PairedPSCBS}
\alias{plotDeltaC1C2}
\alias{plotDeltaC1C2.PairedPSCBS}
\alias{pointsDeltaC1C2}
\alias{pointsDeltaC1C2.PairedPSCBS}
\alias{callABandHighAI}
\alias{callABandHighAI.PairedPSCBS}
\alias{callAllelicBalance}
\alias{callAllelicBalance.PairedPSCBS}
\alias{callExtremeAllelicImbalance}
\alias{callExtremeAllelicImbalance.PairedPSCBS}
\alias{postsegmentTCN}
\alias{postsegmentTCN.PairedPSCBS}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
