%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{PSCBS-package}
\alias{PSCBS-package}
\alias{PSCBS}
\docType{package}

\title{Package PSCBS}


 \description{
   Segments allele-specific DNA copy number data to detect regions
with abnormal copy number within each parental chromosome.

   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires external packages
   R.methodsS3 (>= 1.2.1), R.utils (>= 1.7.8), DNAcopy (>= 1.24.0),
   and also suggests DNAcopy (>= 1.24.0), aroma.light (>= 1.20.0), matrixStats (>= 0.2.2), R.oo (>= 1.8.1), Hmisc (>= 3.8-3).
 }

 \section{Installation and updates}{
   To install this package, use \code{install.packages("PSCBS")}.
 }

 \section{To get started}{
   To get started, see:
   \enumerate{
     \item \code{\link{segmentByPairedPSCBS}}()
   }
 }

 \section{How to cite this package}{
   Please use [1] to cite this package.
 }

 \author{
  A.B. Olshen, H. Bengtsson, P. Neuvial, P.T. Spellman, R.A. Olshen, V.E. Seshan.
 }

 \section{License}{
  GPL (>= 2).
 }

 \references{
  [1] A.B. Olshen, H. Bengtsson, P. Neuvial, P.T. Spellman, R.A. Olshen, V.E. Seshan, \emph{Parent-specific copy number in paired tumor-normal studies using circular binary segmentation}, Bioinformatics 2011 (in press)\cr
 \cr
 }
\keyword{package}
